requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.4
pillow==10.2.0
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
pandas==2.1.4
